add: reuse of settings table
