+++ sources/myapp.js
@@ -1,32 +1,37 @@
 /* eslint-disable no-undef */
-import {JetApp, EmptyRouter, HashRouter, plugins} from "webix-jet";
+import { JetApp, EmptyRouter, HashRouter, plugins } from "webix-jet";
 import "./styles/app.css";
 
 export default class MyApp extends JetApp {
-	constructor(config) {
-		const defaults = {
-			id: APPNAME,
-			version: VERSION,
-			router: BUILD_AS_MODULE ? EmptyRouter : HashRouter,
-			debug: !PRODUCTION,
-			start: "/top/contacts",
-			views: {
-				info: "contacts.Info",
-				form: "contacts.Form"
-			}
-		};
+  constructor(config) {
+    const defaults = {
+      id: APPNAME,
+      version: VERSION,
+      router: BUILD_AS_MODULE ? EmptyRouter : HashRouter,
+      debug: !PRODUCTION,
+      start: "/top/contacts",
+      views: {
+        info: "contacts.Info",
+        form: "contacts.Form",
+      },
+    };
 
-		super({...defaults, ...config});
-	}
+    super({ ...defaults, ...config });
+  }
 }
 
 if (!BUILD_AS_MODULE) {
-	webix.ready(() => {
-		const app = new MyApp();
-		app.use(plugins.Locale);
-		app.render();
-		app.attachEvent("app:error:resolve", () => {
-			webix.delay(() => app.show("/top/contacts"));
-		});
-	});
+  webix.ready(() => {
+    const app = new MyApp();
+    app.use(plugins.Locale, {
+      webix: {
+        en: "en-US",
+        ru: "ru-RU",
+      },
+    });
+    app.render();
+    app.attachEvent("app:error:resolve", () => {
+      webix.delay(() => app.show("/top/contacts"));
+    });
+  });
 }
